# api/openapi.yaml
openapi: 3.1.0
info:
  title: Tailpath API
  version: 0.5.0
  description: Sparse observer ingest, live topology, and bounded history API.
servers:
  - url: /api/v1
paths:
  /capabilities:
    get:
      operationId: getCapabilities
      summary: Return authenticated observer protocol capabilities.
      responses:
        '200':
          description: Server capabilities available to trusted reporters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerCapabilities'
        '401':
          $ref: '#/components/responses/Problem'
  /reports:
    post:
      operationId: submitReport
      summary: Submit a trusted observer report.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportEnvelope'
      responses:
        '202':
          description: Report accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportReceipt'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
  /topology:
    get:
      operationId: getTopology
      summary: Return the current reconciled topology.
      responses:
        '200':
          description: Current topology.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Topology'
        '401':
          $ref: '#/components/responses/Problem'
  /devices:
    get:
      operationId: getDevices
      summary: Return the optional control-plane device directory.
      responses:
        '200':
          description: Current full directory snapshot and synchronization state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceDirectory'
        '401':
          $ref: '#/components/responses/Problem'
  /events:
    get:
      operationId: streamEvents
      summary: Stream topology invalidations over server-sent events.
      responses:
        '200':
          description: SSE event stream.
          content:
            text/event-stream:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Problem'
  /history/edges/{edgeId}:
    get:
      operationId: getEdgeHistory
      summary: Return recent traffic buckets and path events for an edge.
      parameters:
        - name: edgeId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/HistoryWindow'
        - $ref: '#/components/parameters/IncludeSystemTelemetry'
      responses:
        '200':
          description: Edge history.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EdgeHistory'
        '404':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
  /history/nodes:
    get:
      operationId: getHistoryNodes
      summary: Return canonical nodes with traffic in a fixed history window.
      parameters:
        - $ref: '#/components/parameters/HistoryWindow'
        - $ref: '#/components/parameters/IncludeSystemTelemetry'
      responses:
        '200':
          description: Matching canonical node references.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryNodes'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
  /history/edges:
    get:
      operationId: listHistoryEdges
      summary: Return a keyset-paginated page of logical traffic edges.
      parameters:
        - $ref: '#/components/parameters/HistoryWindow'
        - $ref: '#/components/parameters/IncludeSystemTelemetry'
        - name: nodeId
          in: query
          schema:
            type: string
            minLength: 1
        - name: path
          in: query
          description: Match when the path appeared at any point in the window.
          schema:
            $ref: '#/components/schemas/PathKind'
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Matching logical edge page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryEdgePage'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
components:
  parameters:
    HistoryWindow:
      name: window
      in: query
      required: true
      schema:
        $ref: '#/components/schemas/HistoryWindow'
    IncludeSystemTelemetry:
      name: includeSystemTelemetry
      in: query
      required: false
      description: Include Tailpath control-plane telemetry for explicit diagnostics.
      schema:
        type: boolean
        default: false
  responses:
    Problem:
      description: Request failed.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    ServerCapabilities:
      type: object
      required: [observerProtocolVersions, features]
      properties:
        observerProtocolVersions:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: integer
            minimum: 1
        features:
          type: array
          uniqueItems: true
          items:
            type: string
            minLength: 1
    DeviceDirectory:
      type: object
      required: [sync, devices]
      properties:
        sync:
          $ref: '#/components/schemas/DirectorySyncState'
        devices:
          type: array
          items:
            $ref: '#/components/schemas/DirectoryDevice'
    DirectorySyncState:
      type: object
      required: [status, invalidAddressCount]
      properties:
        status:
          type: string
          enum: [disabled, syncing, healthy, stale]
        lastAttemptAt:
          type: string
          format: date-time
        lastSuccessAt:
          type: string
          format: date-time
        nextRetryAt:
          type: string
          format: date-time
        errorCode:
          type: string
          enum: [unauthorized, forbidden, rate-limited, unavailable, timeout, invalid-response]
        invalidAddressCount:
          type: integer
          minimum: 0
...
    ReportEnvelope:
      type: object
      required: [version, reportId, reporterInstanceId, sequence, collectedAt, kind]
      description: >-
        Normal observer messages contain observers. observer_withdrawal contains
        observer identities without peers. relay_session_update contains relaySessions
        instead; the server rejects envelopes that mix the forms.
      properties:
        version:
          type: integer
          const: 1
        reportId:
          type: string
          format: uuid
        reporterInstanceId:
          type: string
          format: uuid
        sequence:
          type: integer
          format: int64
          minimum: 1
        collectedAt:
          type: string
          format: date-time
        kind:
          $ref: '#/components/schemas/ReportKind'
        observers:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ObserverReport'
        relaySessions:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/RelaySessionObservation'
    ReportKind:
      type: string
      enum: [observer_hello, inventory_update, traffic_sample, observer_heartbeat, observer_withdrawal, relay_session_update]
    ObserverReport:
      type: object
      required: [observer, inventoryGeneration]
      description: >-
        For hello and inventory_update, peers is the observer's complete current
        runtime peer view. For traffic_sample it is sparse and contains only
        peers with business byte deltas. observer_withdrawal forbids peers.
        collectedAt is the time for this observer inside a multi-observer
        envelope; legacy reports may omit it and use the envelope timestamp.
      properties:
        observer:
          $ref: '#/components/schemas/NodeIdentity'
        inventoryGeneration:
          type: string
        collectedAt:
          type: string
          format: date-time
        peers:
          type: array
          items:
            $ref: '#/components/schemas/PeerObservation'
...
    RelaySessionObservation:
      type: object
      required:
        - relay
        - source
        - target
        - sessionId
        - vni
        - sourceToTargetBytes
        - targetToSourceBytes
        - sourceToTargetDelta
        - targetToSourceDelta
        - sampleDurationMs
        - lastActive
      description: >-
        A Peer Relay observer's view of traffic between two Tailnet nodes. The
        logical edge is source-target while relay supplies third-party provenance.
      properties:
        relay:
          $ref: '#/components/schemas/NodeIdentity'
          description: The relay observer; stableNodeId is required by the ingest contract.
        source:
          $ref: '#/components/schemas/RelaySessionClient'
        target:
          $ref: '#/components/schemas/RelaySessionClient'
        sessionId:
          type: string
          minLength: 1
        vni:
          type: integer
          format: int64
          minimum: 0
          maximum: 16777215
        sourceToTargetBytes:
          type: integer
          format: int64
          minimum: 0
        targetToSourceBytes:
          type: integer
          format: int64
          minimum: 0
        sourceToTargetDelta:
          type: integer
          format: int64
          minimum: 0
        targetToSourceDelta:
          type: integer
          format: int64
          minimum: 0
        sampleDurationMs:
          type: integer
          format: int64
          minimum: 1
        lastActive:
          type: string
          format: date-time
...
    IdentityStatus:
      type: string
      enum: [resolved, partial, anonymous, conflict]
...
    PathKind:
      type: string
      enum: [direct, derp, peer_relay, unknown]
    ReportReceipt:
      type: object
      required: [accepted, resyncRequired, controlStableNodeIds, heartbeatIntervalMs]
      properties:
        accepted:
          type: boolean
        resyncRequired:
          type: boolean
        controlStableNodeIds:
          type: array
          items:
            type: string
        heartbeatIntervalMs:
          type: integer
          format: int64
          minimum: 1
    Topology:
      type: object
      required: [generatedAt, nodes, edges, observers]
      properties:
        generatedAt:
          type: string
          format: date-time
        nextChangeAt:
          type: string
          format: date-time
          description: Earliest known lifecycle deadline that can change this snapshot.
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/TopologyNode'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/TopologyEdge'
        observers:
          type: array
          items:
            $ref: '#/components/schemas/ObserverState'
    TopologyNode:
      allOf:
        - $ref: '#/components/schemas/NodeIdentity'
        - type: object
          required: [id, observable, online, lastEvidenceAt, clockSkewed]
          properties:
            id:
              type: string
            observable:
              type: boolean
            online:
              type: boolean
            lastEvidenceAt:
              type: string
              format: date-time
            clockSkewed:
              type: boolean
            identityStatus:
              $ref: '#/components/schemas/IdentityStatus'
            directory:
              $ref: '#/components/schemas/DirectoryEnrichment'
    TopologyEdge:
      type: object
      required: [id, source, target, systemTelemetry, path, state, aToBBytesPerSecond, bToABytesPerSecond, lastActive, observations]
      properties:
        id:
          type: string
        source:
          type: string
        target:
          type: string
        systemTelemetry:
          type: boolean
          description: True when either endpoint is the dedicated Tailpath control identity.
        path:
          $ref: '#/components/schemas/PathObservation'
        state:
          type: string
          enum: [active, recent]
        aToBBytesPerSecond:
          type: number
          format: double
          minimum: 0
        bToABytesPerSecond:
          type: number
          format: double
          minimum: 0
        lastActive:
          type: string
          format: date-time
        observations:
          type: array
          items:
            $ref: '#/components/schemas/ObservationProvenance'
        conflicts:
          type: array
          items:
            $ref: '#/components/schemas/PathObservation'
...
    ObserverState:
      type: object
      required: [id, hostname, online, lastSeen, lastCollectedAt, clockSkewMs, clockSkewed]
      properties:
        id:
          type: string
        hostname:
          type: string
        online:
          type: boolean
        lastSeen:
          type: string
          format: date-time
        lastCollectedAt:
          type: string
          format: date-time
        clockSkewMs:
          type: integer
          format: int64
        clockSkewed:
          type: boolean
    EdgeHistory:
      type: object
      required:
        - edgeId
        - source
        - target
        - systemTelemetry
        - relatedNodes
        - from
        - to
        - bucketDurationMs
        - traffic
        - pathEvents
        - trafficTruncated
        - pathEventsTruncated
      properties:
        edgeId:
          type: string
        source:
          $ref: '#/components/schemas/HistoryNodeReference'
        target:
          $ref: '#/components/schemas/HistoryNodeReference'
        systemTelemetry:
          type: boolean
        relatedNodes:
          type: array
          items:
            $ref: '#/components/schemas/HistoryNodeReference'
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        bucketDurationMs:
          type: integer
          format: int64
          minimum: 1
        lastTrafficAt:
          type: string
          format: date-time
          description: Exact server-received time of the last traffic in the selected window. Omitted when the window is empty.
        traffic:
          type: array
          maxItems: 200
          items:
            $ref: '#/components/schemas/TrafficBucket'
        pathAnchor:
          $ref: '#/components/schemas/PathEvent'
        pathEvents:
          type: array
          maxItems: 500
          items:
            $ref: '#/components/schemas/PathEvent'
        trafficTruncated:
          type: boolean
        pathEventsTruncated:
          type: boolean
    HistoryWindow:
      type: string
      enum: [15m, 1h, 6h, 24h, 7d]
...
    HistoryNodes:
      type: object
      required: [nodes]
      properties:
        nodes:
          type: array
          maxItems: 250
          items:
            $ref: '#/components/schemas/HistoryNodeReference'
...
    HistoryEdgePage:
      type: object
      required: [edges]
      properties:
        edges:
          type: array
          items:
            $ref: '#/components/schemas/HistoryEdgeSummary'
        nextCursor:
          type: string
...
    Problem:
      type: object
      required: [title, status]
      properties:
        title:
          type: string
        status:
          type: integer
        detail:
          type: string